Record deleted attachments & prevent thrown exception
Prevent a thrown exception when a message is edited in any server except for our target server. The bot happens to run in multiple servers, but only requires recording edits for one. Therefore, the code has been written in a way that does not hardcode channel values, but only allows recording in one server. Should change in future update.